requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.7.7
    # via -r requirements.in
attrs==21.4.0
    # via pytest
greenlet==1.1.2
    # via sqlalchemy
iniconfig==1.1.1
    # via pytest
injector==0.19.0
    # via -r requirements.in
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
munch==2.5.0
    # via -r requirements.in
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   munch
    #   sqlalchemy-utils
sqlalchemy==1.4.36
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.2
    # via -r requirements.in
tomli==2.0.1
    # via pytest
